refactor: media helpers
