+++ api/review/upload.py
@@ -1,5 +1,4 @@
 import os
-from typing import Any
 
 from fastapi import Header, Query, Request
 from nxtools import logging
@@ -10,7 +9,7 @@ from ayon_server.api.files import handle_upload
 from ayon_server.entities.version import VersionEntity
 from ayon_server.events import EventStream
 from ayon_server.exceptions import BadRequestException
-from ayon_server.helpers.ffprobe import ffprobe
+from ayon_server.helpers.ffprobe import extract_media_info
 from ayon_server.helpers.project_files import id_to_path
 from ayon_server.lib.postgres import Postgres
 from ayon_server.utils import create_uuid
@@ -19,56 +18,9 @@ from .common import ReviewableAuthor, ReviewableModel, availability_from_video_m
 from .router import router
 
 
-async def extract_media_info(file_path: str) -> dict[str, Any]:
-    """Extracts metadata from a video file."""
-
-    try:
-        probe_data = await ffprobe(file_path)
-    except Exception:
-        return {}
-
-    result: dict[str, Any] = {
-        "probeVersion": 1,
-    }
-
-    for stream in probe_data.get("streams", []):
-        if stream.get("codec_type") == "video":
-            fps_str = stream.get("r_frame_rate")
-            fps_parts = fps_str.split("/")
-            if len(fps_parts) == 2:
-                fps = int(fps_parts[0]) / int(fps_parts[1])
-            else:
-                fps = float(fps_parts[0])
-
-            result.update(
-                {
-                    "videoTrackIndex": stream.get("index"),
-                    "width": stream.get("width"),
-                    "height": stream.get("height"),
-                    "pixelFormat": stream.get("pix_fmt"),
-                    "frameRate": fps,
-                    "duration": stream.get("duration"),
-                    "codec": stream.get("codec_name"),
-                }
-            )
-
-        elif stream.get("codec_type") == "audio":
-            if "audioTracks" not in result:
-                result["audioTracks"] = []
-            result["audioTracks"].append(
-                {
-                    "index": stream.get("index"),
-                    "codec": stream.get("codec_name"),
-                    "sampleRate": stream.get("sample_rate"),
-                    "channels": stream.get("channels"),
-                }
-            )
-
-    return result
-
-
 def check_valid_mime(content_type: str) -> None:
     """Checks if the content type is valid for reviewables."""
+    # TODO: replace with helpers.mimetypes functions
 
     if content_type.lower().startswith("video/"):
         return None
@@ -87,7 +39,7 @@ async def upload_reviewable(
     version_id: VersionID,
     label: str | None = Query(None, description="Label", alias="label"),
     content_type: str = Header(...),
-    x_file_name: str | None = Header(None),
+    x_file_name: str = Header(...),
     x_sender: str | None = Header(None),
 ) -> ReviewableModel:
     """Uploads a reviewable for a given version."""
