+++ ayon_server/api/files.py
@@ -5,6 +5,7 @@ from fastapi import Request, Response
 from starlette.responses import FileResponse
 
 from ayon_server.exceptions import AyonException, BadRequestException, NotFoundException
+from ayon_server.helpers.mimetypes import guess_mime_type
 
 
 async def handle_upload(request: Request, target_path: str) -> int:
@@ -62,38 +63,6 @@ async def handle_download(
         filename=filename,
         content_disposition_type=content_disposition_type,
     )
-
-
-def print_escaped_bytes(byte_data: bytes):
-    decoded_str = byte_data.decode("ascii", "ignore")
-    escaped_str = decoded_str.encode("unicode_escape")
-    print(escaped_str)
-
-
-def guess_mime_type(image_bytes: bytes) -> str | None:
-    """Guess the MIME type of an image from its bytes."""
-    if image_bytes[0:4] == b"\x89PNG":
-        media_type = "image/png"
-    elif image_bytes[0:2] == b"\xff\xd8":
-        media_type = "image/jpeg"
-    elif image_bytes[0:4] == b"<svg":
-        media_type = "image/svg+xml"
-    elif image_bytes[0:2] == b"BM":
-        media_type = "image/bmp"
-    elif image_bytes[0:2] == b"II" or image_bytes[0:2] == b"MM":
-        media_type = "image/tiff"
-    elif image_bytes[0:4] == b"RIFF" and image_bytes[8:12] == b"WEBP":
-        media_type = "image/webp"
-    elif image_bytes[0:4] == b"8BPS":
-        media_type = "image/vnd.adobe.photoshop"
-    elif image_bytes[0:4] == b"GIF8":
-        media_type = "image/gif"
-    else:
-        media_type = None
-        bytes_prefix = image_bytes[:16]
-        print("Unknown image format:")
-        print_escaped_bytes(bytes_prefix)
-    return media_type
 
 
 def image_response_from_bytes(image_bytes: bytes) -> Response:
