--- transcoder/__main__.py
@@ -1,9 +1,0 @@
-import asyncio
-
-
-async def main():
-    pass
-
-
-if __name__ == "__main__":
-    asyncio.run(main())
